Adds disableable metadata admin to Services and UKRegions.
