+++ datahub/event/admin.py
@@ -5,30 +5,8 @@ from django.contrib import admin
 from django.contrib.admin import DateFieldListFilter, helpers
 from django.template.response import TemplateResponse
 
-from datahub.core.admin import BaseModelVersionAdmin
+from datahub.core.admin import BaseModelVersionAdmin, DisabledOnFilter
 from datahub.event.models import Event, EventType, LocationType, Programme
-
-
-class DisabledOnFilter(admin.SimpleListFilter):
-    """This filter allows us to filter values that have disabled_on value."""
-
-    title = 'Is disabled'
-    parameter_name = 'disabled_on'
-
-    def lookups(self, request, model_admin):
-        """Returns parameters."""
-        return (
-            ('yes', 'Yes'),
-            ('no', 'No'),
-        )
-
-    def queryset(self, request, queryset):
-        """Modify query according to filter parameter."""
-        value = self.value()
-        if value is not None:
-            is_disabled = True if value == 'yes' else False
-            return queryset.filter(disabled_on__isnull=is_disabled)
-        return queryset
 
 
 def confirm_action(title, action_message):
