+++ datahub/metadata/admin.py
@@ -1,16 +1,15 @@
 from django.contrib import admin
 
+from datahub.core.admin import DisabledOnFilter
 from . import models
 
 MODELS_TO_REGISTER = (
     models.BusinessType,
     models.Sector,
-    models.UKRegion,
     models.Country,
     models.Title,
     models.Role,
     models.TeamRole,
-    models.Service,
     models.ServiceDeliveryStatus,
     models.Event,
     models.InvestmentType,
@@ -21,6 +20,11 @@ MODELS_TO_REGISTER = (
     models.ReferralSourceWebsite,
     models.InvestmentBusinessActivity,
     models.InvestmentStrategicDriver
+)
+
+MODELS_TO_REGISTER_DISABLEABLE = (
+    models.Service,
+    models.UKRegion,
 )
 
 MODELS_TO_REGISTER_WITH_ORDER = (
@@ -35,18 +39,29 @@ MODELS_TO_REGISTER_WITH_ORDER = (
 class MetadataAdmin(admin.ModelAdmin):
     """Custom Metadata Admin."""
 
-    fields = ('name', )
-    list_display = ('name', )
+    fields = ('name',)
+    list_display = ('name',)
     readonly_fields = ('id',)
     search_fields = ('name', 'pk')
+
+
+@admin.register(*MODELS_TO_REGISTER_DISABLEABLE)
+class DisableableMetadataAdmin(admin.ModelAdmin):
+    """Custom Disableable Metadata Admin."""
+
+    fields = ('name', 'disabled_on',)
+    list_display = ('name', 'disabled_on',)
+    readonly_fields = ('id',)
+    search_fields = ('name', 'pk')
+    list_filter = (DisabledOnFilter,)
 
 
 @admin.register(*MODELS_TO_REGISTER_WITH_ORDER)
 class OrderedMetadataAdmin(admin.ModelAdmin):
     """Admin for ordered metadata models."""
 
-    fields = ('name', 'order', )
-    list_display = ('name', 'order', )
+    fields = ('name', 'order',)
+    list_display = ('name', 'order',)
     readonly_fields = ('id',)
     search_fields = ('name', 'pk')
 
@@ -57,5 +72,5 @@ class TeamAdmin(MetadataAdmin):
 
     fields = ('name', 'country', 'uk_region', 'role')
     list_display = ('name', 'role')
-    list_select_related = ('role', )
+    list_select_related = ('role',)
     search_fields = ('name', 'pk')
